HEEDLS-465 - Moved LearningSolutionController into new folder and cleaned up code.
